backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pydantic>=2.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
openai>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests>=2.31.0
schedule>=1.2.0
